backend/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
psycopg2-binary==2.9.9
pgvector==0.4.2

# Redis & Caching
redis==5.0.1
hiredis==3.3.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.2.0
python-multipart==0.0.6
cryptography>=42.0.0
# Task Queue (for future phases)
celery==5.3.4
flower==2.0.1

# AI/ML Stack (for future phases)
langchain>=0.1.0
langgraph>=0.0.40
langchain-core>=0.1.0
langchain-community>=0.0.20
# sentence-transformers is intentionally excluded from default dev setup
# because it pulls torch/CUDA stacks for local embeddings.
openai>=1.6.1
anthropic>=0.7.8

# Network Engineering Tools (for future phases)
scapy==2.5.0
pyshark==0.6
ciscoconfparse==1.9.41
textfsm==1.1.3
ttp==0.9.4

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
structlog==23.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
ruff==0.1.8
mypy==1.7.1
